scripts/requirements.txt
# Requirements needed for the script to run
requests
pyyaml